Added comments to the code and fixed bugs. Simplified the code to receive a response from the neural network

+++ client/src/components/chatsMenu/Chat.jsx
@@ -5,6 +5,7 @@ import { ChatMenuContext } from "../providers/ChatMenuProvider";
 import { ThemeContext } from "../providers/ThemeProvider";
 import ChatService from "../../services/chatServices";
 
+// Changing the color based on the amount of chat dragging
 const interpolateColor = (x, width) => {
   const percentage = Math.min(100, Math.max(0, (Math.abs(x) / width) * 100));
   const red = Math.min(255, Math.floor(255 * (percentage / 100)));
@@ -13,6 +14,7 @@ const interpolateColor = (x, width) => {
   return `rgb(${red}, ${green}, ${blue})`;
 };
 
+// Chat component
 export default function Chat({ element }) {
   const chatRef = useRef(null);
   const [isDarkMode] = useContext(ThemeContext);
@@ -20,6 +22,7 @@ export default function Chat({ element }) {
   const [, setIsChatMenuOpen] = useContext(ChatMenuContext);
   const [isDelete, setIsDelete] = useState(false);
 
+  // Deleting chat animation processing
   const x = useMotionValue(0);
   const backgroundColor = useTransform(
     x,
@@ -30,27 +33,33 @@ export default function Chat({ element }) {
     ]
   );
 
+  // Chat deleting handler
   const deleteChatHandler = async () => {
     if (chatRef && chatRef.current) {
-      const xRange = chatRef.current.style.transform
-        .split("(")[1]
-        .split("px)")[0];
-      if (xRange < window.innerWidth * -0.35) {
-        await ChatService.deleteChat(element.id)
+      // Getting x value of the chat
+      if (chatRef.current.style.transform) {
+        const xRange = chatRef.current.style.transform
+          .split("(")[1]
+          .split("px)")[0];
 
-          .then((res) => {
-            if (selectedChat === element.id) {
-              selectChat(false);
-            }
-            setIsDelete(true);
-            setTimeout(() => {
-              chatRef.current.style.display = "none";
-            }, 200);
-          })
-
-          .catch((error) => {
-            console.log(error);
-          });
+        // Checking the drag value to delete a chat
+        if (xRange < window.innerWidth * -0.35) {
+          setIsDelete(true);
+          await ChatService.deleteChat(element.id)
+            .then((res) => {
+              if (selectedChat === element.id) {
+                selectChat(false);
+              }
+              setTimeout(() => {
+                if (chatRef && chatRef.current) {
+                  chatRef.current.style.display = "none";
+                }
+              }, 500);
+            })
+            .catch((error) => {
+              console.log(error);
+            });
+        }
       }
     }
   };
@@ -63,10 +72,12 @@ export default function Chat({ element }) {
         hidden: {
           x: 0,
           opacity: 1,
+          outline: "2px solid transparent",
         },
         animate: {
           x: "-100%",
           opacity: 0,
+          outline: "2px solid tomato",
         },
       }}
       ref={chatRef}
